refactor: Rename to _xml since compiler only outputs XML
We will then do a general compilation_engine that outputs the VM
language
